Fixed fxml naming and stuff
