+++ src/GUI/Controller/ManagerViewController.java
@@ -1,80 +1,53 @@
 package GUI.Controller;
 
-import BE.User;
-import GUI.Model.DataModel;
-
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXTextArea;
 import java.net.URL;
 import java.util.ResourceBundle;
-
-import javafx.collections.FXCollections;
-import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Label;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.control.TextArea;
+import javafx.scene.image.ImageView;
 
 
 public class ManagerViewController implements Initializable
 {
 
     @FXML
-    private Label lblUserName;
+    private Label lblName;
     @FXML
-    private JFXButton btnSearch;
+    private Label lblPh;
     @FXML
-    private JFXButton btnCopyEmail;
+    private Label lblEmail;
     @FXML
-    private JFXTextArea txtNotes;
+    private Label lblAddress;
     @FXML
-    private JFXButton btnAddHours;
+    private JFXButton btnNameEdit;
     @FXML
-    private JFXButton btnEditInfo;
+    private Label lblHrsAll;
     @FXML
-    private JFXButton btnLogOut;
+    private Label lblHrsYr;
     @FXML
-    private TableView<User> tblVolunteers;
+    private Label lblHrsMth;
     @FXML
-    private TableColumn<User, String> colName;
+    private Label lblHrsDay;
     @FXML
-    private TableColumn<User, Integer> colPhone;
+    private JFXTextArea JFXTxtAreaNotes;
     @FXML
-    private TableColumn<User, String> colEmail;
+    private JFXButton JFXBtnAccept;
     @FXML
-    private TableColumn<User, String> colGuild;
-
+    private JFXButton JFXBtnCancel;
+    @FXML
+    private ImageView imgVwProfilePic;
+    @FXML
+    private TextArea textAreaGuilds;
+    @FXML
+    private JFXButton JFXBtnUpdatePhoto;
     
-    DataModel dataModel = new DataModel();
-    
-    /**
-     * Initializes the controller class.
-     */
     @Override
     public void initialize(URL url, ResourceBundle rb)
     {
-        setTableProperties();
-        tblVolunteers.setItems(FXCollections.observableArrayList(dataModel.getAllUsers()));
-        //System.out.println(dataModel.getAllUsers().toString());
         
     }    
-    
-    /**
-     * Sets table's properties
-     */
-    private void setTableProperties()
-    {
-        colName.setCellValueFactory(new PropertyValueFactory("name"));
-        colPhone.setCellValueFactory(new PropertyValueFactory("phone"));
-        colEmail.setCellValueFactory(new PropertyValueFactory("email"));
-        colGuild.setCellValueFactory(new PropertyValueFactory("guildId"));
-    }
-
-    @FXML
-    private void onLogOutClicked(ActionEvent event)
-    {
-         tblVolunteers.setItems(FXCollections.observableArrayList(dataModel.getAllUsers()));
-    }
 }
